+++ src/main/kotlin/org/agrfesta/btm/api/model/Translation.kt
@@ -1,0 +1,7 @@
+package org.agrfesta.btm.api.model
+
+data class Translation(
+    val text: String,
+    val language: SupportedLanguage,
+    val nonSemanticPart: String? = null
+)